Data_Base_Write_Server/requirements.txt
uvicorn==0.30.1
fastapi==0.111.1
paho-mqtt==2.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.32
psycopg2==2.9.9
logger==1.4